backend/requirements.txt
fastapi
uvicorn
openai==1.12.0
python-dotenv